pyproject.toml
[project]
name = "dnd-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.116.1",
    "fpdf2>=2.8.5",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "typer>=0.21.0",
    "uvicorn>=0.35.0",
]
